Move md into separate MessageContent component
